fix utils test
